requirements.txt
flask~=3.0.3
pandas==2.2.3
numpy==1.26.4
psycopg==3.1.18
psycopg-binary==3.1.18
matplotlib==3.8.3
python-dotenv==1.0.1